+++ samples/cli/index.js
@@ -1,0 +1,4 @@
+var residue = require('residue');
+var logger = residue.getLogger('sample-app');
+
+logger.info('simple log');