+++ samples/server/app.js
@@ -31,9 +31,10 @@ app.get('/default', function (req, res) {
 });
 
 app.get('/', function (req, res) {
-	defaultLogger.info('info default logger');
-	defaultLogger.verbose(3, 'verbose using default logger');
-    namedFunc();
+	//defaultLogger.info('info default logger');
+	//defaultLogger.verbose(3, 'verbose using default logger');
+    logger.info('%s this is best %d:%s', [1, 2, 3], 123, "test");
+    //namedFunc();
     res.send('Hello World!');
 });